Add files via upload
Keyboard PCB specifically for the ZX81+38.
This keyboard uses 6x6mm buttons and a keyboard overlay specifically designed for the ZX-81, with round holes for the button stems.
Its a single sided PCB with round corners designed to fit into the indentation of the keyboard in the ZX-81 enclosure.
buttons are soldered to the top side (with the copper traces), so that pressing the button does not damage the copper pads. after soldering the buttons the leads that stick though are cut off near the PCB so the bottom will be flat.